swupd: check for Valid instead of Populated headers
Change the function to check for valid instead of populated
headers. Checking if the fields were populated during parse will be
handled in the parse function in a subsequent patch.

With the name change, now it is OK to have a Previous == 0 (this is
produced by Mixer at the moment), but is not valid to have Previous
larger than Version.

Reorganized the tests:
- Separated invalid from valid
- Added textual information to invalid, so it is easier to pinpoint what is failing
- Removed the subtest usage from the valid, since was not needed

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>