swupd: vendor github.com/ulikunitz/xz@v0.5.4
This package contains the code to do XZ compression. Will later be
used by fullfile generation. Tag v0.5.4 refers to the latest release
published for that package. The license for the package is a "BSD
3-clause" type.

Note: the blank import was used to signal 'dep ensure' that it should
download this package.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>