swupd: vendor github.com/dsnet/compress/bzip2
This package contains the code to do bzip2 compression. The Go
standard library only have a decompressor for this type of file.

While the repository has no releases, there's evidence the code is
stable enough. See

    "bzip2: merge into Go standard library"
    https://github.com/dsnet/compress/issues/58

for details. The license for the package is a "BSD 3-clause" type.

Note: the blank import was used to signal 'dep ensure' that it should
download this package.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>